pyproject.toml
[project]
name = "yt-uploader"
version = "0.1.0"
description = "Auto-upload videos from USB drives to YouTube with Telegram notifications"
requires-python = ">=3.11"
dependencies = [
    "google-api-python-client>=2.100",
    "google-auth>=2.23",
    "google-auth-oauthlib>=1.1",
    "pyudev>=0.24",
    "httpx>=0.25",
]

[project.scripts]
yt-uploader = "yt_uploader.__main__:main"
yt-uploader-auth = "yt_uploader.auth:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
